Rename the "helpers" folder to "functions"
