+++ frontend/classes/Senator.ts
@@ -3,7 +3,7 @@ import MajorOffice from "@/types/MajorOffice";
 import Praenomen from "@/types/Praenomen";
 import Colors from "@/data/colors.json";
 import FactionNames from "@/data/factionNames.json";
-import getPraenomenAbbr from "@/helpers/praenomen";
+import getPraenomenAbbr from "@/functions/praenomen";
 
 type Name = 'cornelius' | 'fabius' | 'valerius' | 'julius';
