+++ frontend/pages/games/new.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
 import { useRouter } from 'next/router';
 import { useAuth } from '@/contexts/AuthContext';
-import request from "@/helpers/request"
+import request from "@/functions/request"
 import Button from '@/components/Button';
 import { GetServerSidePropsContext } from 'next';
-import getInitialCookieData from '@/helpers/cookies';
+import getInitialCookieData from '@/functions/cookies';
 
 const NewGamePage = () => {
   const router = useRouter();
